apps/network/pyproject.toml
[tool.poetry]
name = "openmined.gridnetwork"
version = "0.1.0"
description = "PyGrid Network"
authors = ["OpenMined <contact@openmined.org>"]
license = "Apache-2.0"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.7"
gevent = "^20.6.2"
gevent-websocket = "^0.10.1"
gunicorn = "^20.0.4"
Flask = "^1.1.1"
Flask-Sockets = "^0.2.1"
requests = "^2.24.0"
idna = "^2.10"
Flask-SQLAlchemy = "^2.4.4"
Flask-Migrate = "^2.5.3"
sqlalchemy_utils = "^0.36.8"
flask-lambda-python36 = "^0.1.0"
sqlalchemy-aurora-data-api = "^0.2.0"
bcrypt = "^3.2.0"
PyJWT = "^1.7.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
coverage = "^4.5.3"
pre-commit = "^2.6.0"
pytest = "^5.4.3"
isort = "^5.0.5"
docformatter = "^1.3.1"
flake8 = "^3.8.3"
flake8-comprehensions = "^3.2.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
